% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_down_mso365.R
\name{my_down_xlsx_mso365}
\alias{my_down_xlsx_mso365}
\title{my_down_xlsx_mso365}
\usage{
my_down_xlsx_mso365(
  file_mso365 = "",
  folder_mso365 = "",
  local_name = "IN",
  where = getwd()
)
}
\arguments{
\item{file_mso365}{OneDrive (online) name of the file to download. Default: ''.}

\item{folder_mso365}{OneDrive (online) folder for the file to download. Default: ''.}

\item{local_name}{Name of the file after downloading. String. Default: 'IN'.}

\item{where}{Destination folder (local). String. Default: current folder (with 'getwd()').}
}
\value{
Nothing.
}
\description{
Function to download a file from Google Drive, given account and file-ID. Using: 'Microsoft365R' package.
}
