% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_load_table.R
\name{my_load_table}
\alias{my_load_table}
\title{my_load_table}
\usage{
my_load_table(
  file_name = "DF",
  sheet_name = NA,
  where = getwd(),
  isa_xlsx = TRUE,
  isa_rds = FALSE,
  isa_rdata = FALSE,
  isa_csv = FALSE,
  row_names = FALSE,
  fileEncoding = "latin1"
)
}
\arguments{
\item{file_name}{Name of the file to load. String. Default: 'DF'.}

\item{sheet_name}{Name of the sheet in which the data to be loaded are (if NA, the first sheet is chosen). String. Default: NA.}

\item{where}{Origin folder. String. Default: current folder (with 'getwd()').}

\item{isa_xlsx}{From a 'xlsx' file. Boolean. Default: TRUE.}

\item{isa_rds}{From a 'rds' file. Boolean. Default: FALSE.}

\item{isa_rdata}{From a 'RData' file. Boolean. Default: FALSE.}

\item{isa_csv}{From a 'csv' file (standard version). Boolean. Default: FALSE.}

\item{row_names}{To use row names. Boolean. Default: FALSE.}

\item{fileEncoding}{File encoding. String. Default: 'latin1'.}
}
\value{
Data-frame.
}
\description{
Function to load a data-frame from 'xlsx', or 'rds', or 'RData', or 'csv'. Using: 'openxlsx' package.
}
