% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice_number.R
\name{my_nice_number}
\alias{my_nice_number}
\title{my_number_marker}
\usage{
my_nice_number(
  x = NA,
  name = NA,
  limits = c(NA, NA),
  dec = 3,
  plus_sign = TRUE,
  equal_sign = FALSE,
  void_string = "∅",
  cuts = c(NA),
  marks = c(NA),
  below_cuts = FALSE,
  abs = FALSE
)
}
\arguments{
\item{x}{Number. Numeric. Default: NA}

\item{name}{Name for x in resulting text (possibly abbreviation). Default: NA}

\item{limits}{Limits for x in resulting text (minimum and maximum). Vector of numerics. Default: c(NA, NA)}

\item{dec}{Number of decimals in resulting text. Numeric. Default: 3}

\item{plus_sign}{Using sign in resulting text (i.e., '+' or '-'). Boolean. Default: TRUE}

\item{equal_sign}{Using equal sign in resulting text (i.e., '='; added by default when there is a name). Boolean. Default: FALSE}

\item{void_string}{String to be used if x is not numeric in resulting text. Default: '∅'}

\item{cuts}{Vector of limits for markers (cut-offs). Vector of numeric. Default: c(NA)}

\item{marks}{Vector of markers (one more than cut-offs). Vector of characters. Default: c(NA)}

\item{below_cuts}{Mark if x is below cut-offs with TRUE (or if above with FALSE). Boolean. Default: FALSE}

\item{abs}{Use the absolute value of x. Boolean. Default: FALSE}
}
\value{
List (x, text, mark).
}
\description{
Function to add markers to a number.
}
