% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_number_marker.R
\name{my_number_marker}
\alias{my_number_marker}
\title{my_number_marker}
\usage{
my_number_marker(
  x = NA,
  cuts = c(NA),
  marks = c(NA),
  below_cuts = FALSE,
  abs = FALSE
)
}
\arguments{
\item{x}{Number. Numeric. Default: NA}

\item{cuts}{Vector of limits for markers (cut-offs). Vector of numerics. Default: c(NA)}

\item{marks}{Vector of markers (one more than cut-offs). Vector of characters. Default: c(NA)}

\item{below_cuts}{Mark if x is below cut-offs with TRUE (or if above with FALSE). Boolean. Default: FALSE}

\item{abs}{Use the absolute value of x. Boolean. Default: FALSE}
}
\value{
Character.
}
\description{
Function to add markers to a number.
}
