% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_save_table.R
\name{my_save_table}
\alias{my_save_table}
\title{my_save_table}
\usage{
my_save_table(
  what,
  file_name = "DF",
  sheet_name = "DATA",
  add_sheet = FALSE,
  where = getwd(),
  freeze_colums = 1,
  with_rds = TRUE,
  with_rdata = FALSE,
  with_csv = FALSE,
  row_names = FALSE,
  fileEncoding = "latin1"
)
}
\arguments{
\item{what}{Data-frame to be saved. Data-frame. Default: None.}

\item{file_name}{Name by which the file will be saved. String. Default: 'DF'.}

\item{sheet_name}{Name of the sheet in which the file will be saved. String. Default: 'DATA'.}

\item{add_sheet}{To add a sheet in an existing file (instead of create a new file). Boolean. Default: FALSE.}

\item{where}{Destination folder. String. Default: current folder (with 'getwd()').}

\item{freeze_colums}{Number of columns to be freezed. Default: 1.}

\item{with_rds}{To save as 'rds'. Boolean. Default: TRUE.}

\item{with_rdata}{To save as 'RData'. Boolean. Default: FALSE.}

\item{with_csv}{To save as 'csv' (both standard and Italian-version 'csv' are saved). Boolean. Default: FALSE.}

\item{row_names}{To use row names. Boolean. Default: FALSE.}

\item{fileEncoding}{File encoding. String. Default: 'latin1'.}
}
\value{
Nothing.
}
\description{
Function to save a data-frame in 'xlsx', and/or 'rds', and/or 'RData', and/or 'csv'. Using: 'openxlsx' package.
}
