% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_up_gdrive.R
\name{my_up_gdrive}
\alias{my_up_gdrive}
\title{my_up_gdrive}
\usage{
my_up_gdrive(
  account = "",
  current_file = "",
  file_name_uploaded = "NEW",
  where_to_upload = ""
)
}
\arguments{
\item{account}{Account in Google Drive (usually e-mail). Default: ''.}

\item{current_file}{Name of the file to upload. String. Default: ''.}

\item{file_name_uploaded}{Name of the file after uploading (online). String. Default: 'NEW'.}

\item{where_to_upload}{Folder in Gogle Drive (online) in which the file will be uploaded. String. Default: ''}
}
\value{
Nothing.
}
\description{
Function to upload a file in a folder of Google Drive, given account and a local file (overwriting any online file with the same name). Using: 'googledrive' package.
}
