% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_up_mso365.R
\name{my_up_mso365}
\alias{my_up_mso365}
\title{my_up_mso365}
\usage{
my_up_mso365(
  current_file = "",
  file_name_uploaded = "NEW",
  where_to_upload = ""
)
}
\arguments{
\item{current_file}{Name of the file to upload. String. Default: ''.}

\item{file_name_uploaded}{Name of the file after uploading. String. Default: 'NEW'.}

\item{where_to_upload}{Folder in Gogle Drive in which the file will be uploaded. String. Default: ''}
}
\value{
Nothing.
}
\description{
Function to download a file from Google Drive, given account and file-ID. Using: 'Microsoft365R' package.
}
